Create utilitary function for operation definition
